+++ src/main/java/com/previnci/Util/ConfigReader.java
@@ -18,7 +18,7 @@ public class ConfigReader {
 
         prop = new Properties();
         try {
-            FileInputStream ip = new FileInputStream("./src/test/resources/config/config.properties");
+            FileInputStream ip = new FileInputStream("./src/test/resources/config.properties");
             prop.load(ip);
 
         } catch (FileNotFoundException e) {
